Switched from lua_gpio library to rpio library and did code cleanup
switched the gpio library due to incompatability and lack of support for
lua_gpio, the new libary needed some tweeking but it now works way
better then the old one. Also cleaned up and reworked some of the code,
switching buttons from triggering on a high pin to trigging on a low
pin.